+++ index.html
@@ -6,7 +6,7 @@
     <link rel="icon" href="/favicon.ico">
     <title></title>
   </head>
-  <body class="bg-gradient-to-b from-green-300 to-purple-500">
+  <body class="bg-gradient-to-b from-transparent to-sky-200">
     <noscript>
       <strong>We're sorry but this site doesn't work properly without JavaScript enabled. Please enable it to continue.</strong>
     </noscript>
